[OneExplorer] Move _nodeMap to OneStorage
This commit refactors OneExplorer by relocating _nodeMap into OneStorage.

ONE-vscode-DCO-1.0-Signed-off-by: Dayoung Lee <dayoung.lee@samsung.com>
